# src/main/resources/openapi.spec.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Carol's App Prototype",
    "description" : "Prototype of Carols app, to begin to prove some concepts and gather data. React frontend. Java backend. Postgres datastore. Running on Heroku.",
    "termsOfService" : "None",
    "contact" : {
      "email" : "oliver.hulett@gmail.com"
    },
    "version" : "1.0"
  },
  "paths" : {
    "/ingredient/create-or-update" : {
      "post" : {
        "operationId" : "createOrUpdateIngredients",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Ingredient"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ingredient/delete" : {
      "post" : {
        "operationId" : "deleteIngredientsById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/ingredient/get" : {
      "post" : {
        "operationId" : "getIngredientsById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/create-or-update" : {
      "post" : {
        "operationId" : "createOrUpdateProducts",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/get" : {
      "post" : {
        "operationId" : "getProductsById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/delete" : {
      "post" : {
        "operationId" : "deleteProductsById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/supplier/create-or-update" : {
      "post" : {
        "operationId" : "createOrUpdateSuppliers",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Supplier"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Supplier"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/supplier/get" : {
      "post" : {
        "operationId" : "getSuppliersById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Supplier"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/supplier/delete" : {
      "post" : {
        "operationId" : "deleteSuppliersById",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Ingredient" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "thumbnail" : {
            "type" : "string"
          }
        }
      },
      "Product" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "thumbnail" : {
            "type" : "string"
          }
        }
      },
      "Supplier" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "thumbnail" : {
            "type" : "string"
          }
        }
      }
    }
  }
}